+++ backend/src/main/java/com/empire/World.java
@@ -2,36 +2,24 @@ package com.empire;
 
 import com.google.appengine.api.datastore.Blob;
 import com.google.appengine.api.datastore.DatastoreService;
-import com.google.appengine.api.datastore.DatastoreServiceFactory;
-import com.google.appengine.api.datastore.EmbeddedEntity;
-import com.google.appengine.api.datastore.Entity;
 import com.google.appengine.api.datastore.Entity;
 import com.google.appengine.api.datastore.EntityNotFoundException;
-import com.google.appengine.api.datastore.Key;
 import com.google.appengine.api.datastore.KeyFactory;
 import com.google.appengine.api.datastore.Text;
 import com.google.common.primitives.Ints;
 import com.google.gson.FieldNamingPolicy;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
-import java.io.IOException;
-import java.nio.charset.StandardCharsets;
-import java.security.SecureRandom;
 import java.time.DayOfWeek;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.Comparator;
 import java.util.HashMap;
 import java.util.HashSet;
-import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
-import java.util.PriorityQueue;
 import java.util.Set;
 import java.util.TreeSet;
-import java.util.function.Function;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
@@ -82,7 +70,7 @@ class World implements GoodwillProvider {
 		return getGson().fromJson(json, World.class);
 	}
 
-	public static World startNew(String gmPasswordHash, String obsPasswordHash, Map<String, Nation.NationGson> nationSetup) {
+	public static World startNew(String gmPasswordHash, String obsPasswordHash, Map<String, Nation> nationSetup) {
 		World w = new World();
 		w.date = 1;
 		w.gmPasswordHash = gmPasswordHash;
@@ -128,7 +116,7 @@ class World implements GoodwillProvider {
 		double totalSharesFood = unruledNations / 2.0;
 		double totalSharesPopulation = unruledNations / 2.0;
 		for (String kingdom : nationSetup.keySet()) {
-			Nation.NationGson setup = nationSetup.get(kingdom);
+			Nation setup = nationSetup.get(kingdom);
 			totalSharesGold += 1;
 			totalSharesArmy += 1;
 			totalSharesNavy += 1;
@@ -148,7 +136,7 @@ class World implements GoodwillProvider {
 			else if ("gold".equals(setup.bonus)) totalSharesGold += 0.5;
 		}
 		for (String kingdom : nationSetup.keySet()) {
-			Nation.NationGson setup = nationSetup.get(kingdom);
+			Nation setup = nationSetup.get(kingdom);
 			NationData nation = new NationData();
 			WorldConstantData.Kingdom con = WorldConstantData.kingdoms.get(kingdom);
 			nation.email = setup.email;
@@ -219,7 +207,7 @@ class World implements GoodwillProvider {
 		// Un-own rebellious regions.
 		HashSet<String> rebelliousNations = new HashSet<>();
 		for (String kingdom : nationSetup.keySet()) {
-			Nation.NationGson setup = nationSetup.get(kingdom);
+			Nation setup = nationSetup.get(kingdom);
 			if (!"Rebellious".equals(setup.trait1) && !"Rebellious".equals(setup.trait2)) continue;
 			rebelliousNations.add(kingdom);
 			ArrayList<Region> ownedRegions = new ArrayList<>();
@@ -236,7 +224,7 @@ class World implements GoodwillProvider {
 		}
 		// Allocate nobles as necessary.
 		for (String kingdom : nationSetup.keySet()) {
-			Nation.NationGson setup = nationSetup.get(kingdom);
+			Nation setup = nationSetup.get(kingdom);
 			Culture culture = WorldConstantData.kingdoms.get(kingdom).culture;
 			if (setup.hasTag(NationData.Tag.REPUBLICAN)) continue;
 			ArrayList<Integer> ownedRegions = new ArrayList<>();
@@ -293,7 +281,7 @@ class World implements GoodwillProvider {
 		}
 		// Place armies, navies.
 		for (String kingdom : nationSetup.keySet()) {
-			Nation.NationGson setup = nationSetup.get(kingdom);
+			Nation setup = nationSetup.get(kingdom);
 			double sharesNavy = 1;
 			double sharesArmy = 1;
 			if (setup.hasTag(NationData.Tag.PATRIOTIC)) sharesArmy += 0.15;
@@ -364,7 +352,7 @@ class World implements GoodwillProvider {
 		}
 		// Add characters, incl Cardinals
 		for (String kingdom : nationSetup.keySet()) {
-			Nation.NationGson setup = nationSetup.get(kingdom);
+			Nation setup = nationSetup.get(kingdom);
 			ArrayList<Integer> regions = new ArrayList<>();
 			for (int i = 0; i < w.regions.size(); i++) if (kingdom.equals(w.regions.get(i).getKingdom())) regions.add(i);
 			log.log(Level.INFO, "Setting up " + kingdom + ", " + regions.size());
